requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
pymongo==4.5.0
flask-jwt-extended==4.4.4
Werkzeug==3.0.0
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
